redone auth components & custom loading component
